Explicit set Int64 in test for 32bit environment

+++ test/misc.jl
@@ -481,7 +481,7 @@ end
 @test Base.prettyprint_getunits(1023, 6, 1024)[2] == 1
 @test Base.prettyprint_getunits(1024, 6, 1024) == (1, 2)
 @test Base.prettyprint_getunits(1025, 6, 1024)[2] == 2
-@test Base.prettyprint_getunits(1024^5 - 1, 6, 1024)[2] == 5
-@test Base.prettyprint_getunits(1024^5, 6, 1024) == (1, 6)
-@test Base.prettyprint_getunits(1024^5 + 1, 6, 1024)[2] == 6
-@test Base.prettyprint_getunits(1024^5, 5, 1024) == (1024, 5)
+@test Base.prettyprint_getunits(Int64(1024)^5 - 1, 6, 1024)[2] == 5
+@test Base.prettyprint_getunits(Int64(1024)^5, 6, 1024) == (1, 6)
+@test Base.prettyprint_getunits(Int64(1024)^5 + 1, 6, 1024)[2] == 6
+@test Base.prettyprint_getunits(Int64(1024)^5, 5, 1024) == (1024, 5)
